refactor code for workspaces
