+++ asg_rev/users/serializers/user.py
@@ -1,5 +1,5 @@
 from rest_framework import serializers
-from users.models.user import User
+from users.models import User
 
 class UserSerializer(serializers.ModelSerializer):
     password = serializers.CharField(write_only=True)
